CurrentScene component
