+++ apps/desktop/src/renderer/components/CurrentScene/CurrentScene.module.css
@@ -1,0 +1,10 @@
+.wrapper {
+  display: flex;
+  flex-direction: column;
+  height: 100%;
+}
+
+.content {
+  flex: 1;
+  min-height: 0;
+}
