+++ apps/desktop/src/renderer/components/WorkArea/WorkArea.tsx
@@ -1,9 +1,9 @@
 import { Intro } from '@components/Intro/Intro'
-import { Sketches } from '@components/Sketches/Sketches'
+import { CurrentScene } from '@components/CurrentScene/CurrentScene'
 import { useAppStore } from '@renderer/appStore'
 
 export const WorkArea = () => {
   const isShowingSketches = useAppStore((state) => state.sketchesDir)
 
-  return isShowingSketches ? <Sketches /> : <Intro />
+  return isShowingSketches ? <CurrentScene /> : <Intro />
 }
